Create DomainToLegacyConverter
