+++ src/DataAccess/LegacyConverters/DomainToLegacyConverter.php
@@ -1,0 +1,150 @@
+<?php
+
+declare( strict_types = 1 );
+
+namespace WMDE\Fundraising\MembershipContext\DataAccess\LegacyConverters;
+
+use Doctrine\Common\Collections\ArrayCollection;
+use Doctrine\Common\Collections\Collection;
+use WMDE\Fundraising\MembershipContext\DataAccess\DoctrineEntities\MembershipApplication as DoctrineApplication;
+use WMDE\Fundraising\MembershipContext\DataAccess\Exception\UnknownIncentive;
+use WMDE\Fundraising\MembershipContext\DataAccess\MembershipApplicationData;
+use WMDE\Fundraising\MembershipContext\Domain\Model\Applicant;
+use WMDE\Fundraising\MembershipContext\Domain\Model\MembershipApplication;
+use WMDE\Fundraising\MembershipContext\Domain\Model\Payment;
+use WMDE\Fundraising\PaymentContext\Domain\Model\BankData;
+use WMDE\Fundraising\PaymentContext\Domain\Model\DirectDebitPayment;
+use WMDE\Fundraising\PaymentContext\Domain\Model\PaymentMethod;
+use WMDE\Fundraising\PaymentContext\Domain\Model\PayPalData;
+use WMDE\Fundraising\PaymentContext\Domain\Model\PayPalPayment;
+
+class DomainToLegacyConverter {
+	public function convert( DoctrineApplication $doctrineApplication, MembershipApplication $application, IncentiveFinder $finder ): void {
+		$doctrineApplication->setId( $application->getId() );
+		$doctrineApplication->setMembershipType( $application->getType() );
+
+		$this->setApplicantFields( $doctrineApplication, $application->getApplicant() );
+		$this->setPaymentFields( $doctrineApplication, $application->getPayment() );
+		$doctrineApplication->setIncentives( $this->convertIncentives( $finder, $application->getIncentives() ) );
+		$doctrineApplication->setDonationReceipt( $application->getDonationReceipt() );
+
+		$doctrineStatus = $this->getDoctrineStatus( $application );
+		$this->preserveDoctrineStatus( $doctrineApplication, $doctrineStatus );
+		$doctrineApplication->setStatus( $doctrineStatus );
+	}
+
+	private function convertIncentives( IncentiveFinder $finder, \Traversable $incentives ): Collection {
+		$incentiveCollection = new ArrayCollection();
+		foreach ( $incentives as $incentive ) {
+			if ( $incentive->getId() !== null ) {
+				$incentiveCollection->add( $incentive );
+				continue;
+			}
+			$foundIncentive = $finder->findIncentiveByName( $incentive->getName() );
+			if ( $foundIncentive === null ) {
+				throw new UnknownIncentive( sprintf( 'Incentive "%s" not found', $incentive->getName() ) );
+			}
+			$incentiveCollection->add( $foundIncentive );
+		}
+		return $incentiveCollection;
+	}
+
+	private function setApplicantFields( DoctrineApplication $application, Applicant $applicant ): void {
+		$application->setApplicantFirstName( $applicant->getName()->getFirstName() );
+		$application->setApplicantLastName( $applicant->getName()->getLastName() );
+		$application->setApplicantSalutation( $applicant->getName()->getSalutation() );
+		$application->setApplicantTitle( $applicant->getName()->getTitle() );
+		$application->setCompany( $applicant->getName()->getCompanyName() );
+
+		$application->setApplicantDateOfBirth( $applicant->getDateOfBirth() );
+
+		$application->setApplicantEmailAddress( $applicant->getEmailAddress()->getFullAddress() );
+		$application->setApplicantPhoneNumber( $applicant->getPhoneNumber()->__toString() );
+
+		$address = $applicant->getPhysicalAddress();
+
+		$application->setCity( $address->getCity() );
+		$application->setCountry( $address->getCountryCode() );
+		$application->setPostcode( $address->getPostalCode() );
+		$application->setAddress( $address->getStreetAddress() );
+	}
+
+	private function setPaymentFields( DoctrineApplication $application, Payment $payment ): void {
+		$application->setPaymentIntervalInMonths( $payment->getIntervalInMonths() );
+		$application->setPaymentAmount( (int)$payment->getAmount()->getEuroFloat() );
+		$paymentMethod = $payment->getPaymentMethod();
+
+		$application->setPaymentType( $paymentMethod->getId() );
+		if ( $paymentMethod instanceof DirectDebitPayment ) {
+			$this->setBankDataFields( $application, $paymentMethod->getBankData() );
+		} elseif ( $paymentMethod instanceof PayPalPayment && $paymentMethod->getPayPalData() != new PayPalData() ) {
+			$this->setPayPalDataFields( $application, $paymentMethod->getPayPalData() );
+		}
+	}
+
+	private function setBankDataFields( DoctrineApplication $application, BankData $bankData ): void {
+		$application->setPaymentBankAccount( $bankData->getAccount() );
+		$application->setPaymentBankCode( $bankData->getBankCode() );
+		$application->setPaymentBankName( $bankData->getBankName() );
+		$application->setPaymentBic( $bankData->getBic() );
+		$application->setPaymentIban( $bankData->getIban()->toString() );
+	}
+
+	private function setPayPalDataFields( DoctrineApplication $application, PayPalData $payPalData ): void {
+		$application->encodeAndSetData( array_merge(
+			$application->getDecodedData(),
+			[
+				'paypal_payer_id' => $payPalData->getPayerId(),
+				'paypal_subscr_id' => $payPalData->getSubscriberId(),
+				'paypal_payer_status' => $payPalData->getPayerStatus(),
+				'paypal_address_status' => $payPalData->getAddressStatus(),
+				'paypal_mc_gross' => $payPalData->getAmount()->getEuroString(),
+				'paypal_mc_currency' => $payPalData->getCurrencyCode(),
+				'paypal_mc_fee' => $payPalData->getFee()->getEuroString(),
+				'paypal_settle_amount' => $payPalData->getSettleAmount()->getEuroString(),
+				'paypal_first_name' => $payPalData->getFirstName(),
+				'paypal_last_name' => $payPalData->getLastName(),
+				'paypal_address_name' => $payPalData->getAddressName(),
+				'ext_payment_id' => $payPalData->getPaymentId(),
+				'ext_subscr_id' => $payPalData->getSubscriberId(),
+				'ext_payment_type' => $payPalData->getPaymentType(),
+				'ext_payment_status' => $payPalData->getPaymentStatus(),
+				'ext_payment_account' => $payPalData->getPayerId(),
+				'ext_payment_timestamp' => $payPalData->getPaymentTimestamp(),
+				'first_payment_date' => $payPalData->getFirstPaymentDate()
+			]
+		) );
+	}
+
+	private function getDoctrineStatus( MembershipApplication $application ): int {
+		if ( $application->needsModeration() && $application->isCancelled() ) {
+			return DoctrineApplication::STATUS_CANCELLED_MODERATION;
+		}
+
+		if ( $application->needsModeration() ) {
+			return DoctrineApplication::STATUS_MODERATION;
+		}
+
+		if ( $application->isCancelled() ) {
+			return DoctrineApplication::STATUS_CANCELED;
+		}
+
+		if ( !$application->getPayment()->getPaymentMethod()->hasExternalProvider() ) {
+			return DoctrineApplication::STATUS_CONFIRMED;
+		}
+
+		if ( $application->getPayment()->getPaymentMethod()->paymentCompleted() ) {
+			return DoctrineApplication::STATUS_CONFIRMED;
+		}
+
+		return DoctrineApplication::STATUS_NEUTRAL;
+	}
+
+	private function preserveDoctrineStatus( DoctrineApplication $doctrineApplication, int $doctrineStatus ): void {
+		if ( $doctrineStatus < DoctrineApplication::STATUS_CONFIRMED ) {
+			$doctrineApplication->modifyDataObject( function ( MembershipApplicationData $data ): void {
+				$data->setPreservedStatus( DoctrineApplication::STATUS_CONFIRMED );
+			} );
+		}
+	}
+}
