+++ src/DataAccess/LegacyConverters/IncentiveFinder.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace WMDE\Fundraising\MembershipContext\DataAccess\LegacyConverters;
+
+use WMDE\Fundraising\MembershipContext\Domain\Model\Incentive;
+
+interface IncentiveFinder {
+	public function findIncentiveByName( string $name ): ?Incentive;
+}
